+++ Runtime/MonoBehaviourExtensions.cs
@@ -8,23 +8,19 @@ namespace AncientLightStudios.Nanoject.MonoBehaviours
     {
 
         /// <summary>
-        /// Scans the current scene for a MonoBehaviour that derives from the given type and declares it in the
-        /// dependency context. The MonoBehaviour will not be qualified. If the scan yields
-        /// multiple objects of the given type, an <see cref="InvalidOperationException"/> will be thrown as you
-        /// can only have one unqualified object of the same type in a dependency context.
+        /// Scans the current scene for MonoBehaviours that derive from the given type and declares them in the
+        /// dependency context. The MonoBehaviours will not be qualified.
         /// </summary>
-        /// <param name="context">the context into which the instance should be declared.</param>
+        /// <param name="context">the context into which the instances should be declared.</param>
         /// </param>
-        public static void DeclareMonoBehaviourFromScene<T>(this DependencyContext context) where T : MonoBehaviour
+        public static void DeclareMonoBehavioursFromScene<T>(this DependencyContext context) where T : MonoBehaviour
         {
             DeclareMonoBehavioursFromScene<T>(context, false);
         }
 
         /// <summary>
         /// Scans the current scene for MonoBehaviours that derive from the given type and declares them in the
-        /// dependency context. The MonoBehaviours will be qualified using their name in the scene hierarchy. If the
-        /// scan yields multiple objects of the given type with the same name, an <see cref="InvalidOperationException"/>
-        /// will be thrown as you can only have one qualified object of the same type and name in a dependency context.
+        /// dependency context. The MonoBehaviours will be qualified using their name in the scene hierarchy.
         /// </summary>
         /// <param name="context">the context into which the classes should be declared.</param>
         /// </param>
@@ -39,9 +35,7 @@ namespace AncientLightStudios.Nanoject.MonoBehaviours
         /// </summary>
         /// <param name="context">the context into which the classes should be declared.</param>
         /// <param name="qualify">if true, the scanned component will be declared with their name as qualifier.
-        /// If false, they will be declared without a qualifier. Note that when this is false and the scan yields
-        /// multiple objects of the given type, an <see cref="InvalidOperationException"/> will be thrown as you
-        /// can only have one unqualified object of the same type in a dependency context.
+        /// If false, they will be declared without a qualifier. 
         /// </param>
         private static void DeclareMonoBehavioursFromScene<T>(DependencyContext context, bool qualify) where T:MonoBehaviour
         {
@@ -54,18 +48,6 @@ namespace AncientLightStudios.Nanoject.MonoBehaviours
                 if (qualify)
                 {
                     qualifier = result.name;
-                }
-
-                if (context.IsDeclared(type, qualifier))
-                {
-                    if (qualifier == "")
-                    {
-                        throw new InvalidOperationException( $"Found multiple objects of type {type.Name} with no qualifier.");
-                    }
-                    else
-                    {
-                        throw new InvalidOperationException( $"Found multiple objects of type {type.Name} with the same qualifier '{qualifier}'.");
-                    }
                 }
                     
                 context.DeclareQualified(type, qualifier, result);